+++ SearchResult.cpp
@@ -97,7 +97,7 @@ string SearchResult::getSlotString() const {
 }
 
 int64_t SearchResult::getConnectionInt() const {
-	return isNMDC() ? Util::toDouble(connection)*1024.0*1024.0/8.0 : Util::toInt64(connection);
+	return isNMDC() ? static_cast<int64_t>(Util::toDouble(connection)*1024.0*1024.0/8.0) : Util::toInt64(connection);
 }
 
 int64_t SearchResult::getSpeedPerSlot() const {
